feat(rmcp/reorg): reorganize code structure to accomodate v1.5 and v2.0 better
