+++ src/connection/impls/rmcp/asf.rs
@@ -1,5 +1,3 @@
-use super::encapsulation::{CalculateAuthCodeError, IpmiSessionMessage, UnwrapEncapsulationError};
-
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct SupportedInteractions {
     pub rcmp_security: bool,
@@ -138,7 +136,7 @@ pub struct ASFMessage {
 }
 
 impl ASFMessage {
-    fn write_data(&self, buffer: &mut Vec<u8>) {
+    pub(crate) fn write_data(&self, buffer: &mut Vec<u8>) {
         buffer.extend_from_slice(&4542u32.to_le_bytes());
 
         buffer.push(self.message_type.type_byte());
@@ -148,7 +146,7 @@ impl ASFMessage {
         self.message_type.write_data(buffer);
     }
 
-    fn from_bytes(data: &[u8]) -> Option<Self> {
+    pub(crate) fn from_bytes(data: &[u8]) -> Option<Self> {
         if data.len() < 8 {
             return None;
         }
@@ -169,123 +167,6 @@ impl ASFMessage {
         Some(Self {
             message_tag,
             message_type,
-        })
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum RmcpClass {
-    Ack(u8),
-    Asf(ASFMessage),
-    Ipmi(IpmiSessionMessage),
-    OemDefined,
-}
-
-impl RmcpClass {
-    fn write_data(
-        &self,
-        password: Option<&[u8; 16]>,
-        buffer: &mut Vec<u8>,
-    ) -> Result<(), CalculateAuthCodeError> {
-        match self {
-            RmcpClass::Ack(_) => {
-                log::trace!("Received RMCP ACK, but do not know how to validate sequence number.");
-                Ok(())
-            }
-            RmcpClass::Asf(message) => {
-                message.write_data(buffer);
-                Ok(())
-            }
-            RmcpClass::Ipmi(message) => message.write_data(password, buffer),
-            // TODO: OEMDefined data
-            RmcpClass::OemDefined => todo!(),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum RmcpUnwrapError {
-    /// There was not enough data in the packet to parse a valid RMCP message.
-    NotEnoughData,
-    /// An error occurred while trying to unwrap the encapsulated RMCP
-    /// message.
-    UnwrapEncapsulation(UnwrapEncapsulationError),
-    /// The RMCP packet contained an invalid ASF message.
-    InvalidASFMessage,
-    /// The class of the RMCP packet was not valid.
-    InvalidRmcpClass,
-}
-
-impl From<UnwrapEncapsulationError> for RmcpUnwrapError {
-    fn from(value: UnwrapEncapsulationError) -> Self {
-        Self::UnwrapEncapsulation(value)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct RmcpMessage {
-    pub version: u8,
-    pub sequence_number: u8,
-    pub class_and_contents: RmcpClass,
-}
-
-impl RmcpMessage {
-    pub fn new(sequence_number: u8, contents: RmcpClass) -> Self {
-        Self {
-            version: 6,
-            sequence_number,
-            class_and_contents: contents,
-        }
-    }
-
-    pub fn to_bytes(&self, password: Option<&[u8; 16]>) -> Result<Vec<u8>, CalculateAuthCodeError> {
-        let class = match self.class_and_contents {
-            RmcpClass::Ack(value) => value | 0x80,
-            RmcpClass::Asf(_) => 0x06,
-            RmcpClass::Ipmi(_) => 0x07,
-            RmcpClass::OemDefined => 0x08,
-        };
-
-        let sequence_number = if matches!(self.class_and_contents, RmcpClass::Ipmi(_)) {
-            0xFF
-        } else {
-            self.sequence_number
-        };
-
-        let mut bytes = vec![self.version, 0, sequence_number, class];
-
-        self.class_and_contents.write_data(password, &mut bytes)?;
-
-        Ok(bytes)
-    }
-
-    pub fn from_bytes(password: Option<&[u8; 16]>, data: &[u8]) -> Result<Self, RmcpUnwrapError> {
-        if data.len() < 4 {
-            return Err(RmcpUnwrapError::NotEnoughData);
-        }
-
-        let version = data[0];
-        let sequence_number = data[2];
-        let class = data[3];
-
-        let data = &data[4..];
-
-        let class = match class {
-            0x06 => RmcpClass::Asf(
-                ASFMessage::from_bytes(data).ok_or(RmcpUnwrapError::InvalidASFMessage)?,
-            ),
-            0x07 => RmcpClass::Ipmi(IpmiSessionMessage::from_bytes(data, password)?),
-            0x08 => RmcpClass::OemDefined,
-            _ if class & 0x80 == 0x80 => RmcpClass::Ack(class & 0x7F),
-            _ => {
-                return Err(RmcpUnwrapError::InvalidRmcpClass);
-            }
-        };
-
-        Ok(Self {
-            version,
-            sequence_number,
-            class_and_contents: class,
         })
     }
 }
