+++ src/connection/impls/rmcp/v2_0/authentication.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum AuthenticationAlgorithm {
+    RakpNone,
+    RakpHmacSha1,
+    RakpHmacMd5,
+    RakpHmacSha256,
+    Oem(u8),
+}
