+++ src/connection/impls/rmcp/v2_0/confidentiality.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum ConfidentialityAlgorithm {
+    AesCbc128,
+    Xrc4_128,
+    Xrc4_40,
+    Oem(u8),
+}
